@@ -116,8 +116,8 @@ Notifications.unregister = function() {
 /**
  * Local Notifications
  * @param {Object}		details
+ * @param {String}		details.title  -  The title displayed in the notification alert.
  * @param {String}		details.message - The message displayed in the notification alert.
- * @param {String}		details.title  -  ANDROID ONLY: The title displayed in the notification alert.
  * @param {String}		details.ticker -  ANDROID ONLY: The ticker displayed in the status bar.
  * @param {Object}		details.userInfo -  iOS ONLY: The userInfo used in the notification alert.
  */
@@ -131,7 +131,7 @@ Notifications.localNotification = function(details: Object) {
 			soundName = ''; // empty string results in no sound (and no vibration)
 		}
 
-		// for valid fields see: https://developer.apple.com/library/ios/documentation/NetworkingInternet/Conceptual/RemoteNotificationsPG/Chapters/IPhoneOSClientImp.html
+		// for valid fields see: https://developer.apple.com/library/archive/documentation/NetworkingInternet/Conceptual/RemoteNotificationsPG/PayloadKeyReference.html
 		// alertTitle only valid for apple watch: https://developer.apple.com/library/ios/documentation/iPhone/Reference/UILocalNotification_Class/#//apple_ref/occ/instp/UILocalNotification/alertTitle
 
 		this.handler.presentLocalNotification({
@@ -163,10 +163,13 @@ Notifications.localNotificationSchedule = function(details: Object) {
 
 		const iosDetails = {
 			fireDate: details.date.toISOString(),
+			alertTitle: details.title,
 			alertBody: details.message,
+			category: details.category,
 			soundName: soundName,
 			userInfo: details.userInfo,
-			repeatInterval: details.repeatType
+			repeatInterval: details.repeatType,
+			category: details.category,
 		};
 
 		if(details.number) {
@@ -182,7 +185,7 @@ Notifications.localNotificationSchedule = function(details: Object) {
 		details.fireDate = details.date.getTime();
 		delete details.date;
 		// ignore iOS only repeatType
-		if (['year', 'month'].includes(details.repeatType)) {
+		if (['year'].includes(details.repeatType)) {
 			delete details.repeatType;
 		}
 		this.handler.scheduleLocalNotification(details);
@@ -274,6 +277,10 @@ Notifications.requestPermissions = function() {
 };
 
 /* Fallback functions */
+Notifications.subscribeToTopic = function() {
+	return this.callNative('subscribeToTopic', arguments);
+};
+
 Notifications.presentLocalNotification = function() {
 	return this.callNative('presentLocalNotification', arguments);
 };
@@ -284,6 +291,10 @@ Notifications.scheduleLocalNotification = function() {
 
 Notifications.cancelLocalNotifications = function() {
 	return this.callNative('cancelLocalNotifications', arguments);
+};
+
+Notifications.clearLocalNotification = function() {
+    return this.callNative('clearLocalNotification', arguments);
 };
 
 Notifications.cancelAllLocalNotifications = function() {
